SFOS-428 Eclipse whines about a lot of minor things
spelling corrections and remove some unneeded casts

+++ trunk/core/smartfrog/src/org/smartfrog/examples/dynamicwebserver/gui/graphpanel/GraphImpl.java
@@ -186,7 +186,7 @@ public class GraphImpl extends CompoundImpl implements Graph, Compound,
         histogram = sfResolve(HISTOGRAM, false, false);
         pencilColour = sfResolve(PENCILCOLOUR, "0xFFFF00", false);
         positionDisplay = sfResolve(POSITIONDISPLAY, "NW", false);
-        graphPencilWidth = (float) sfResolve(GRAPHPENCILWIDTH, 5, false);
+        graphPencilWidth = sfResolve(GRAPHPENCILWIDTH, 5, false);
         frameTitle = sfResolve(FRAMETITLE, "Graph", false);
         logger = new LogWrapper((Logger) sfResolve(LOGTO, false));
     }
